Starting to phase-in the new parser. I moved 'Dao.Token.Location' to the "Dao.NewParser" module. The old 'Dao.Token.Location' is no longer used anywhere throughout the program. The old parser still makes use of the "Dao.Token" module. It is worth noting that the reason "Dao.Token" was separate from "Dao.Parser" was because of 'Dao.Token.Location' which was used for not just the abstract syntax tree, but "Dao.Object" expressions as well. However, the "Dao.Parser" module imports "Dao.Object" so importing "Dao.Parser" to make use of the 'Dao.Object.Location' would result in a circular dependency. The "Dao.NewParser" does not make use of "Dao.Object" and so can be imported safely in most all other modules.

+++ src/Dao/Evaluator.hs
@@ -28,7 +28,8 @@
 module Dao.Evaluator where
 
 import           Dao.Debug.OFF
-import           Dao.Token
+-- import           Dao.Token
+import           Dao.NewParser
 import           Dao.Object
 import           Dao.Object.AST
 import           Dao.PPrint
@@ -1157,8 +1158,8 @@ lookupFunction msg op = do
 errAt :: Location -> [Object]
 errAt loc = case loc of
   LocationUnknown -> []
-  loc -> [ OPair (OWord (startingLine loc), OWord (fromIntegral (startingColumn loc)))
-         , OPair (OWord (endingLine   loc), OWord (fromIntegral (endingColumn   loc)))
+  loc -> [ OPair (OWord (fromIntegral (startingLine loc)), OWord (fromIntegral (startingColumn loc)))
+         , OPair (OWord (fromIntegral (endingLine   loc)), OWord (fromIntegral (endingColumn   loc)))
          ]
 
 -- | Evaluate to 'procErr' if the given 'PValue' is 'Backtrack' or 'PFail'. You must pass a
